requirements.txt
flask
fabric
gunicorn
gevent
redis
docker-py
celery
